[PS-78] refactor: 사용자 로그인/회원가입 로직을 AuthService로 이동
