+++ src/main/java/com/tf4/photospot/auth/application/AuthService.java
@@ -4,8 +4,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.tf4.photospot.auth.application.response.ReissueTokenResponse;
-import com.tf4.photospot.user.application.UserService;
+import com.tf4.photospot.global.exception.ApiException;
+import com.tf4.photospot.global.exception.domain.UserErrorCode;
+import com.tf4.photospot.user.application.request.LoginUserInfo;
+import com.tf4.photospot.user.application.response.OauthLoginUserResponse;
 import com.tf4.photospot.user.domain.User;
+import com.tf4.photospot.user.domain.UserRepository;
+import com.tf4.photospot.user.util.NicknameGenerator;
 
 import lombok.RequiredArgsConstructor;
 
@@ -14,12 +19,31 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class AuthService {
 
-	private final UserService userService;
+	private final UserRepository userRepository;
 	private final JwtService jwtService;
 
 	public ReissueTokenResponse reissueToken(Long userId, String refreshToken) {
 		jwtService.validRefreshToken(userId, refreshToken);
-		User user = userService.findUser(userId);
+		User user = userRepository.findById(userId).orElseThrow(() -> new ApiException(UserErrorCode.NOT_FOUND_USER));
 		return new ReissueTokenResponse(jwtService.issueAccessToken(user.getId(), user.getRole().type));
+	}
+
+	public OauthLoginUserResponse oauthLogin(String providerType, String account) {
+		return userRepository.findUserByProviderTypeAndAccount(providerType, account)
+			.map(findUser -> OauthLoginUserResponse.from(true, findUser))
+			.orElseGet(() -> OauthLoginUserResponse.from(false,
+				userRepository.save(new LoginUserInfo(providerType, account).toUser(generateNickname()))));
+	}
+
+	private boolean isNicknameDuplicated(String nickname) {
+		return userRepository.existsByNickname(nickname);
+	}
+
+	private String generateNickname() {
+		String generatedRandomNickname = NicknameGenerator.generatorRandomNickname();
+		while (isNicknameDuplicated(generatedRandomNickname)) {
+			generatedRandomNickname = NicknameGenerator.generatorRandomNickname();
+		}
+		return generatedRandomNickname;
 	}
 }
