+++ src/test/java/com/tf4/photospot/auth/application/AuthServiceTest.java
@@ -7,14 +7,12 @@ import org.junit.jupiter.api.Test;
 
 import com.tf4.photospot.auth.application.response.ReissueTokenResponse;
 import com.tf4.photospot.support.IntegrationTestSupport;
-import com.tf4.photospot.user.application.UserService;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
 public class AuthServiceTest extends IntegrationTestSupport {
 	private final AuthService authService;
-	private final UserService userService;
 	private final JwtService jwtService;
 
 	@DisplayName("DB에 저장된 리프레시 토큰과 전달 받은 토큰이 같으면 액세스 토큰을 재발급 한다.")
@@ -23,7 +21,7 @@ public class AuthServiceTest extends IntegrationTestSupport {
 		// given
 		String providerType = "kakao";
 		String account = "account";
-		Long loginUser = userService.oauthLogin(providerType, account).getId();
+		Long loginUser = authService.oauthLogin(providerType, account).getId();
 		String refreshToken = jwtService.issueRefreshToken(loginUser);
 
 		// when
